+++ R/prep_cyclocomp.R
@@ -3,13 +3,7 @@
 #' @importFrom cyclocomp cyclocomp_package_dir
 
 PREPS$cyclocomp <- function(state, path = state$path, quiet) {
-  state$cyclocomp <- try(cyclocomp_package_dir(path), silent = quiet)
-  if(inherits(state$cyclocomp, "try-error")) {
-    warning(
-      "Prep step for cyclocomp failed: ",
-      conditionMessage(attr(state$cyclocomp, "condition")),
-      call. = FALSE
-    )
-  }
-  state
+  run_prep_step(state, "cyclocomp", function() {
+    cyclocomp_package_dir(path)
+  }, quiet = quiet)
 }