+++ R/prep_description.R
@@ -3,14 +3,7 @@
 #' @importFrom desc description
 
 PREPS$description <- function(state, path = state$path, quiet) {
-  state$description <- try(description$new(file.path(path, "DESCRIPTION")), 
-                           silent = quiet)
-  if(inherits(state$description, "try-error")) {
-    warning(
-      "Prep step for description failed: ",
-      conditionMessage(attr(state$description, "condition")),
-      call. = FALSE
-    )
-  }
-  state
+  run_prep_step(state, "description", function() {
+    description$new(file.path(path, "DESCRIPTION"))
+  }, quiet = quiet)
 }