+++ R/prep_lintr.R
@@ -24,16 +24,7 @@ linters_to_lint <- list(
 
 PREPS$lintr <- function(state, path = state$path, quiet) {
   path <- normalizePath(path)
-  suppressMessages(
-    state$lintr <- try(lint_package(path, linters = linters_to_lint),
-                       silent = TRUE)
-  )
-  if(inherits(state$lintr, "try-error")) {
-    warning(
-      "Prep step for lintr failed: ",
-      conditionMessage(attr(state$lintr, "condition")),
-      call. = FALSE
-    )
-  }
-  state
+  run_prep_step(state, "lintr", function() {
+    suppressMessages(lint_package(path, linters = linters_to_lint))
+  }, quiet = quiet)
 }