+++ R/prep_rcmdcheck.R
@@ -4,13 +4,7 @@
 
 PREPS$rcmdcheck <- function(state, path = state$path, quiet) {
   path <- normalizePath(path)
-  state$rcmdcheck <- try(rcmdcheck(path, quiet = quiet), silent = quiet)
-  if(inherits(state$rcmdcheck, "try-error")) {
-    warning(
-      "Prep step for rcmdcheck failed: ",
-      conditionMessage(attr(state$rcmdcheck, "condition")),
-      call. = FALSE
-    )
-  }
-  state
+  run_prep_step(state, "rcmdcheck", function() {
+    rcmdcheck(path, quiet = quiet)
+  }, quiet = quiet)
 }